:sparkles: feat: 将动作库设置为固定
