+++ src/store/setting/selectors/config.ts
@@ -13,10 +13,6 @@ const currentTouchConfig = (s: SettingStore): TouchConfig => {
   return s.config.touch;
 };
 
-const getAnimationsByGender = (s: SettingStore, gender: GenderEnum) => {
-  return s.config.animation?.[gender] || [];
-};
-
 const getTouchActionsByGenderAndArea = (
   s: SettingStore,
   gender: GenderEnum,
@@ -29,6 +25,5 @@ const getTouchActionsByGenderAndArea = (
 export const configSelectors = {
   currentOpenAIConfig,
   currentTouchConfig,
-  getAnimationsByGender,
   getTouchActionsByGenderAndArea,
 };
